susceptible transitions
